Move form inside its own component
